Added texting, email, sound notifications plus general polishing
Texting notifications not in a fully functional state. Need more work around frequent status code 403 around the post requests on the login stage. Possibly a loop around having a proxy list and multiple user agents may solve the issue. Sound files are currently being hosted in the root jupyter user folder which are exposed to the users. If contained in hidden folder, the sound files can't be accessed. May be a permissions issue.